requirements.txt
wget
pandas
csv
json
psycopg2
sqlalchemy
pymongow